Folder and imports renaming.
